+++ LittleGame/EnemyAttackComponent.h
@@ -29,9 +29,9 @@ public:
 	2. Determines this with an internal timer which compares with attackDuration
 	*/
 	virtual bool timeToAttack() {
-		this->passedTime += Locator::getGameTime()->getDeltaTime();
+		this->passedTime += static_cast<float>(Locator::getGameTime()->getDeltaTime());
 		if (this->passedTime > attackDuration) {
-			this->passedTime = 0;
+			this->passedTime = 0.0f;
 			return true;
 		}
 		return false;
